Fixed types (c_ulong -> c_uint)

+++ Engine/plugins/macro.py
@@ -34,15 +34,15 @@ MALWARE_ID_OLE = 99
 # -------------------------------------------------------------------------
 BYTE = c_ubyte
 WORD = c_ushort
-DWORD = c_ulong
+DWORD = c_uint
 FLOAT = c_float
 LPBYTE = POINTER(c_ubyte)
 LPTSTR = POINTER(c_char)
 HANDLE = c_void_p
 PVOID = c_void_p
 LPVOID = c_void_p
-UINT_PTR = c_ulong
-SIZE_T = c_ulong
+UINT_PTR = c_uint
+SIZE_T = c_uint
 
 
 class MCD(Structure):
